Moved factory into a separate entity package.
