+++ backend/src/main/java/entity/recurrent_monthly_income_factory/CommonRecurrentMonthlyIncomeFactory.java
@@ -1,8 +1,10 @@
-package entity.monthly_income;
+package entity.recurrent_monthly_income_factory;
 
 import entity.item_income.CommonItemIncomeFactory;
 import entity.item_income.ItemIncome;
 import entity.item_income.ItemIncomeFactory;
+import entity.monthly_income.CommonMonthlyIncome;
+import entity.monthly_income.MonthlyIncome;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -11,6 +13,9 @@ public class CommonRecurrentMonthlyIncomeFactory implements RecurrentMonthlyInco
 
     private final List<ItemIncome> recurrentItems = new ArrayList<>();
     private final ItemIncomeFactory itemIncomeFactory = new CommonItemIncomeFactory();
+
+
+    public CommonRecurrentMonthlyIncomeFactory() {}
 
     @Override
     public MonthlyIncome create(String date) {
